+++ app/src/main/java/cc/kafuu/bilidownload/bilibili/video/ResourceDownloadCallback.java
@@ -3,7 +3,7 @@ package cc.kafuu.bilidownload.bilibili.video;
 import java.io.File;
 
 public interface ResourceDownloadCallback {
-    void onStatus(int current, int contentLength);
+    void onStatus(long current, long contentLength);
     void onStop();
     void onCompleted(File file);
     void onFailure(String message);
